Crate libR_sys

A low-level libR binding library which is kept deliberately minimal.

In particular, it has no external dependencies other that libR installed on the target.

The libR-sys crate is a low level bindgen wrapper for the R programming language. The intention is to allow one or more extension mechanisms to be implemented for rust.

Effort to make the extension libraries platform-independent can be concentrated here.

Examples

use libR_sys::{Rf_initialize_R, R_CStackLimit, setup_Rmainloop};
use std::os::raw;
 
unsafe {
  std::env::set_var("R_HOME", "/usr/lib/R");
  let arg0 = "R\0".as_ptr() as *mut raw::c_char;
  Rf_initialize_R(1, [arg0].as_mut_ptr());
  R_CStackLimit = usize::max_value();
  setup_Rmainloop();
}

Structs

R_allocator

R_inpstream_st

R_long_vec_hdr_t

R_outpstream_st

Rcomplex

SEXPREC

_G_fpos64_t

_G_fpos_t

_IO_FILE

_IO_FILE_plus

_IO_jump_t

_IO_marker

__fsid_t

__mbstate_t

__va_list_tag

exception

max_align_t

structRstart

Constants

AIO_PRIO_DELTA_MAX

ANYSXP

ARG_MAX

BCODESXP

BC_BASE_MAX

BC_DIM_MAX

BC_SCALE_MAX

BC_STRING_MAX

BUFSIZ

BUILTINSXP

CHARCLASS_NAME_MAX

CHARSXP

CLOSXP

COLL_WEIGHTS_MAX

CPLXSXP

DELAYTIMER_MAX

DOMAIN

DOTSXP

ENABLE_NLS

ENVSXP

EOF

EXPRSXP

EXPR_NEST_MAX

EXTPTRSXP

FILENAME_MAX

FOPEN_MAX

FP_ILOGB0

FP_ILOGBNAN

FP_INFINITE

FP_NAN

FP_NORMAL

FP_SUBNORMAL

FP_ZERO

FREESXP

FUNSXP

HAVE_ALLOCA_H

HAVE_F77_UNDERSCORE

HAVE_VISIBILITY_ATTRIBUTE

HOST_NAME_MAX

IEEE_754

INTSXP

LANGSXP

LGLSXP

LINE_MAX

LINK_MAX

LISTSXP

LOGIN_NAME_MAX

L_ctermid

L_tmpnam

MATH_ERREXCEPT

MATH_ERRNO

MAX_CANON

MAX_INPUT

MAX_NUM_SEXPTYPE

MB_LEN_MAX

MQ_PRIO_MAX

M_1_PI

M_2_PI

M_2_SQRTPI

M_E

M_LN2

M_LN10

M_LOG2E

M_LOG10E

M_PI

M_PI_2

M_PI_4

M_SQRT2

M_SQRT1_2

NAMEDMAX

NAME_MAX

NEWSXP

NGROUPS_MAX

NILSXP

NR_OPEN

OVERFLOW

PATH_MAX

PIPE_BUF

PLOSS

PROMSXP

PTHREAD_DESTRUCTOR_ITERATIONS

PTHREAD_KEYS_MAX

PTHREAD_STACK_MIN

P_tmpdir

ParseStatus_PARSE_EOF

ParseStatus_PARSE_ERROR

ParseStatus_PARSE_INCOMPLETE

ParseStatus_PARSE_NULL

ParseStatus_PARSE_OK

RAWSXP

REALSXP

RE_DUP_MAX

RTSIG_MAX

R_LONG_VEC_TOKEN

R_SHORT_LEN_MAX

R_XDR_DOUBLE_SIZE

R_XDR_INTEGER_SIZE

R_XLEN_T_MAX

R_pstream_format_t_R_pstream_any_format

R_pstream_format_t_R_pstream_ascii_format

R_pstream_format_t_R_pstream_asciihex_format

R_pstream_format_t_R_pstream_binary_format

R_pstream_format_t_R_pstream_xdr_format

Rboolean_FALSE

Rboolean_TRUE

S4SXP

SA_TYPE_SA_DEFAULT

SA_TYPE_SA_NORESTORE

SA_TYPE_SA_NOSAVE

SA_TYPE_SA_RESTORE

SA_TYPE_SA_SAVE

SA_TYPE_SA_SAVEASK

SA_TYPE_SA_SUICIDE

SEEK_CUR

SEEK_END

SEEK_SET

SEM_VALUE_MAX

SING

SIZEOF_SIZE_T

SPECIALSXP

STRSXP

SUPPORT_MBCS

SUPPORT_OPENMP

SUPPORT_UTF8

SYMSXP

TLOSS

TMP_MAX

TTY_NAME_MAX

TYPE_BITS

UNDERFLOW

VECSXP

WEAKREFSXP

XATTR_LIST_MAX

XATTR_NAME_MAX

XATTR_SIZE_MAX

_ATFILE_SOURCE

_BITS_LIBM_SIMD_DECL_STUBS_H

_BITS_POSIX1_LIM_H

_BITS_POSIX2_LIM_H

_BITS_TYPESIZES_H

_BITS_TYPES_H

_DEFAULT_SOURCE

_FEATURES_H

_G_BUFSIZ

_G_HAVE_MMAP

_G_HAVE_MREMAP

_G_IO_IO_FILE_VERSION

_G_config_h

_IOFBF

_IOLBF

_IONBF

_IOS_APPEND

_IOS_ATEND

_IOS_BIN

_IOS_INPUT

_IOS_NOCREATE

_IOS_NOREPLACE

_IOS_OUTPUT

_IOS_TRUNC

_IO_BAD_SEEN

_IO_BOOLALPHA

_IO_BUFSIZ

_IO_CURRENTLY_PUTTING

_IO_DEC

_IO_DELETE_DONT_CLOSE

_IO_DONT_CLOSE

_IO_EOF_SEEN

_IO_ERR_SEEN

_IO_FIXED

_IO_FLAGS2_MMAP

_IO_FLAGS2_NOTCANCEL

_IO_FLAGS2_USER_WBUF

_IO_HEX

_IO_INTERNAL

_IO_IN_BACKUP

_IO_IS_APPENDING

_IO_IS_FILEBUF

_IO_LEFT

_IO_LINE_BUF

_IO_LINKED

_IO_MAGIC

_IO_MAGIC_MASK

_IO_NO_READS

_IO_NO_WRITES

_IO_OCT

_IO_RIGHT

_IO_SCIENTIFIC

_IO_SHOWBASE

_IO_SHOWPOINT

_IO_SHOWPOS

_IO_SKIPWS

_IO_STDIO

_IO_TIED_PUT_GET

_IO_UNBUFFERED

_IO_UNIFIED_JUMPTABLES

_IO_UNITBUF

_IO_UPPERCASE

_IO_USER_BUF

_IO_USER_LOCK

_LIBC_LIMITS_H_

_LIB_VERSION_TYPE__IEEE_

_LIB_VERSION_TYPE__ISOC_

_LIB_VERSION_TYPE__POSIX_

_LIB_VERSION_TYPE__SVID_

_LIB_VERSION_TYPE__XOPEN_

_MATH_H

_MATH_H_MATHDEF

_OLD_STDIO_MAGIC

_POSIX2_BC_BASE_MAX

_POSIX2_BC_DIM_MAX

_POSIX2_BC_SCALE_MAX

_POSIX2_BC_STRING_MAX

_POSIX2_COLL_WEIGHTS_MAX

_POSIX2_EXPR_NEST_MAX

_POSIX2_LINE_MAX

_POSIX2_RE_DUP_MAX

_POSIX2_CHARCLASS_NAME_MAX

_POSIX_AIO_LISTIO_MAX

_POSIX_AIO_MAX

_POSIX_ARG_MAX

_POSIX_CHILD_MAX

_POSIX_CLOCKRES_MIN

_POSIX_C_SOURCE

_POSIX_DELAYTIMER_MAX

_POSIX_HOST_NAME_MAX

_POSIX_LINK_MAX

_POSIX_LOGIN_NAME_MAX

_POSIX_MAX_CANON

_POSIX_MAX_INPUT

_POSIX_MQ_OPEN_MAX

_POSIX_MQ_PRIO_MAX

_POSIX_NAME_MAX

_POSIX_NGROUPS_MAX

_POSIX_OPEN_MAX

_POSIX_PATH_MAX

_POSIX_PIPE_BUF

_POSIX_RE_DUP_MAX

_POSIX_RTSIG_MAX

_POSIX_SEM_NSEMS_MAX

_POSIX_SEM_VALUE_MAX

_POSIX_SIGQUEUE_MAX

_POSIX_SOURCE

_POSIX_SSIZE_MAX

_POSIX_STREAM_MAX

_POSIX_SYMLINK_MAX

_POSIX_SYMLOOP_MAX

_POSIX_THREAD_DESTRUCTOR_ITERATIONS

_POSIX_THREAD_KEYS_MAX

_POSIX_THREAD_THREADS_MAX

_POSIX_TIMER_MAX

_POSIX_TTY_NAME_MAX

_POSIX_TZNAME_MAX

_STDC_PREDEF_H

_STDIO_H

_SYS_CDEFS_H

__FD_SETSIZE

__FILE_defined

__GLIBC_MINOR__

__GLIBC__

__GNUC_VA_LIST

__GNU_LIBRARY__

__INO_T_MATCHES_INO64_T

__MATH_DECLARE_LDOUBLE

__MATH_DECLARING_DOUBLE

__OFF_T_MATCHES_OFF64_T

__STDC_IEC_559__

__STDC_IEC_559_COMPLEX__

__STDC_ISO_10646__

__STDC_NO_THREADS__

__SYSCALL_WORDSIZE

__USE_ATFILE

__USE_FORTIFY_LEVEL

__USE_ISOC11

__USE_ISOC95

__USE_ISOC99

__USE_MISC

__USE_POSIX

__USE_POSIX2

__USE_POSIX199309

__USE_POSIX199506

__USE_POSIX_IMPLICITLY

__USE_XOPEN2K

__USE_XOPEN2K8

__WORDSIZE

__WORDSIZE_TIME64_COMPAT32

____FILE_defined

____mbstate_t_defined

__codecvt_result___codecvt_error

__codecvt_result___codecvt_noconv

__codecvt_result___codecvt_ok

__codecvt_result___codecvt_partial

cetype_t_CE_ANY

cetype_t_CE_BYTES

cetype_t_CE_LATIN1

cetype_t_CE_NATIVE

cetype_t_CE_SYMBOL

cetype_t_CE_UTF8

math_errhandling

nchar_type_Bytes

nchar_type_Chars

nchar_type_Width

Statics

R_BaseEnv

R_BaseNamespace

R_BaseSymbol

R_BlankScalarString

R_BlankString

R_BraceSymbol

R_Bracket2Symbol

R_BracketSymbol

R_CStackLimit

R_CStackStart

R_ClassSymbol

R_Consolefile

R_DeviceSymbol

R_DimNamesSymbol

R_DimSymbol

R_DirtyImage

R_DollarSymbol

R_DotsSymbol

R_DoubleColonSymbol

R_DropSymbol

R_EmptyEnv

R_GUIType

R_GlobalContext

R_GlobalEnv

R_HistoryFile

R_HistorySize

R_Home

R_Interactive

R_LastvalueSymbol

R_LevelsSymbol

R_MissingArg

R_ModeSymbol

R_NaInt

R_NaN

R_NaReal

R_NaRmSymbol

R_NaString

R_NameSymbol

R_NamesSymbol

R_NamespaceEnvSymbol

R_NamespaceRegistry

R_NegInf

R_NilValue

R_Outputfile

R_PackageSymbol

R_PosInf

R_PreviousSymbol

R_QuoteSymbol

R_RestartToken

R_RestoreHistory

R_RowNamesSymbol

R_SeedsSymbol

R_SignalHandlers

R_Slave

R_SortListSymbol

R_SourceSymbol

R_SpecSymbol

R_Srcref

R_TempDir

R_TripleColonSymbol

R_TspSymbol

R_UnboundValue

R_baseSymbol

R_dot_Method

R_dot_defined

R_dot_packageName

R_dot_target

R_running_as_main_program

_IO_2_1_stdin_

_IO_2_1_stdout_

_IO_2_1_stderr_

_LIB_VERSION

signgam

stderr

stdin

stdout

sys_errlist

sys_nerr

Functions

ATTRIB⚠

BODY⚠

CAAR⚠

CAD4R⚠

CADDDR⚠

CADDR⚠

CADR⚠

CAR⚠

CDAR⚠

CDDDR⚠

CDDR⚠

CDR⚠

CLOENV⚠

COMPLEX⚠

CONS_NR⚠

CleanEd⚠

DDVAL⚠

DUPLICATE_ATTRIB⚠

ENCLOS⚠

ENVFLAGS⚠

FORMALS⚠

FRAME⚠

HASHASH⚠

HASHTAB⚠

HASHVALUE⚠

INTEGER⚠

INTERNAL⚠

IS_LONG_VEC⚠

IS_S4_OBJECT⚠

LENGTH⚠

LEVELS⚠

LOGICAL⚠

MARK⚠

MISSING⚠

NAMED⚠

OBJECT⚠

PRCODE⚠

PRENV⚠

PRINTNAME⚠

PRSEEN⚠

PRVALUE⚠

RAW⚠

RDEBUG⚠

REAL⚠

REFCNT⚠

REprintf⚠

REvprintf⚠

RSTEP⚠

RTRACE⚠

R_BindingIsActive⚠

R_BindingIsLocked⚠

R_CHAR⚠

R_CheckStack⚠

R_CheckStack2⚠

R_CheckUserInterrupt⚠

R_CleanTempDir⚠

R_ClearExternalPtr⚠

R_ClearerrConsole⚠

R_ClosureExpr⚠

R_DefParams⚠

R_EnvironmentIsLocked⚠

R_ExecWithCleanup⚠

R_ExpandFileName⚠

R_ExternalPtrAddr⚠

R_ExternalPtrProtected⚠

R_ExternalPtrTag⚠

R_FindNamespace⚠

R_FindPackageEnv⚠

R_FixupRHS⚠

R_FlushConsole⚠

R_GetCurrentSrcref⚠

R_GetSrcFilename⚠

R_HasFancyBindings⚠

R_HomeDir⚠

R_InitFileInPStream⚠

R_InitFileOutPStream⚠

R_InitInPStream⚠

R_InitOutPStream⚠

R_IsNA⚠

R_IsNaN⚠

R_IsNamespaceEnv⚠

R_IsPackageEnv⚠

R_LockBinding⚠

R_LockEnvironment⚠

R_MakeActiveBinding⚠

R_MakeExternalPtr⚠

R_MakeWeakRef⚠

R_MakeWeakRefC⚠

R_NamespaceEnvSpec⚠

R_PackageEnvName⚠

R_ParseVector⚠

R_PreserveObject⚠

R_PromiseExpr⚠

R_ProtectWithIndex⚠

R_RegisterCFinalizer⚠

R_RegisterCFinalizerEx⚠

R_RegisterFinalizer⚠

R_RegisterFinalizerEx⚠

R_ReleaseObject⚠

R_ReplDLLdo1⚠

R_ReplDLLinit⚠

R_Reprotect⚠

R_RestoreGlobalEnv⚠

R_RestoreGlobalEnvFromFile⚠

R_RestoreHashCount⚠

R_RunExitFinalizers⚠

R_RunPendingFinalizers⚠

R_RunWeakRefFinalizer⚠

R_SaveGlobalEnv⚠

R_SaveGlobalEnvToFile⚠

R_Serialize⚠

R_SetExternalPtrAddr⚠

R_SetExternalPtrProtected⚠

R_SetExternalPtrTag⚠

R_SetParams⚠

R_SetWin32⚠

R_ShowMessage⚠

R_SizeFromEnv⚠

R_Suicide⚠

R_ToplevelExec⚠

R_Unserialize⚠

R_WeakRefKey⚠

R_WeakRefValue⚠

R_XDRDecodeDouble⚠

R_XDRDecodeInteger⚠

R_XDREncodeDouble⚠

R_XDREncodeInteger⚠

R_alloc⚠

R_allocLD⚠

R_atof⚠

R_bcDecode⚠

R_bcEncode⚠

R_check_class_and_super⚠

R_check_class_etc⚠

R_common_command_line⚠

R_compute_identical⚠

R_csort⚠

R_curErrorBuf⚠

R_cycle_detected⚠

R_do_MAKE_CLASS⚠

R_do_new_object⚠

R_do_slot⚠

R_do_slot_assign⚠

R_dot_Last⚠

R_envHasNoSpecialSymbols⚠

R_extends⚠

R_finite⚠

R_forceAndCall⚠

R_gc⚠

R_gc_running⚠

R_getClassDef⚠

R_getClassDef_R⚠

R_has_methods_attached⚠

R_has_slot⚠

R_initialize_bcode⚠

R_isVirtualClass⚠

R_isort⚠

R_lsInternal⚠

R_lsInternal3⚠

R_max_col⚠

R_nchar⚠

R_orderVector⚠

R_qsort⚠

R_qsort_I⚠

R_qsort_int⚠

R_qsort_int_I⚠

R_rsort⚠

R_setStartTime⚠

R_set_command_line_arguments⚠

R_setupHistory⚠

R_signal_protect_error⚠

R_signal_reprotect_error⚠

R_signal_unprotect_error⚠

R_strtod⚠

R_system⚠

R_tmpnam⚠

R_tmpnam2⚠

R_tryEval⚠

R_tryEvalSilent⚠

R_unLockBinding⚠

Rf_CreateTag⚠

Rf_DropDims⚠

Rf_GetArrayDimnames⚠

Rf_GetColNames⚠

Rf_GetMatrixDimnames⚠

Rf_GetOption⚠

Rf_GetOption1⚠

Rf_GetOptionDigits⚠

Rf_GetOptionWidth⚠

Rf_GetRowNames⚠

Rf_KillAllDevices⚠

Rf_NonNullStringMatch⚠

Rf_PairToVectorList⚠

Rf_PrintValue⚠

Rf_S3Class⚠

Rf_ScalarComplex⚠

Rf_ScalarInteger⚠

Rf_ScalarLogical⚠

Rf_ScalarRaw⚠

Rf_ScalarReal⚠

Rf_ScalarString⚠

Rf_StringBlank⚠

Rf_StringFalse⚠

Rf_StringTrue⚠

Rf_VectorToPairList⚠

Rf_acopy_string⚠

Rf_addMissingVarsToNewEnv⚠

Rf_alloc3DArray⚠

Rf_allocArray⚠

Rf_allocFormalsList2⚠

Rf_allocFormalsList3⚠

Rf_allocFormalsList4⚠

Rf_allocFormalsList5⚠

Rf_allocFormalsList6⚠

Rf_allocList⚠

Rf_allocMatrix⚠

Rf_allocS4Object⚠

Rf_allocSExp⚠

Rf_allocVector⚠

Rf_allocVector3⚠

Rf_any_duplicated⚠

Rf_any_duplicated3⚠

Rf_applyClosure⚠

Rf_arraySubscript⚠

Rf_asChar⚠

Rf_asCharacterFactor⚠

Rf_asComplex⚠

Rf_asInteger⚠

Rf_asLogical⚠

Rf_asReal⚠

Rf_asS4⚠

Rf_cPsort⚠

Rf_classgets⚠

Rf_coerceVector⚠

Rf_conformable⚠

Rf_cons⚠

Rf_copyListMatrix⚠

Rf_copyMatrix⚠

Rf_copyMostAttrib⚠

Rf_copyVector⚠

Rf_countContexts⚠

Rf_defineVar⚠

Rf_dimgets⚠

Rf_dimnamesgets⚠

Rf_duplicate⚠

Rf_duplicated⚠

Rf_elt⚠

Rf_endEmbeddedR⚠

Rf_error⚠

Rf_errorcall⚠

Rf_eval⚠

Rf_findFun⚠

Rf_findVar⚠

Rf_findVarInFrame⚠

Rf_findVarInFrame3⚠

Rf_getAttrib⚠

Rf_getCharCE⚠

Rf_gsetVar⚠

Rf_iPsort⚠

Rf_inherits⚠

Rf_initEmbeddedR⚠

Rf_initialize_R⚠

Rf_install⚠

Rf_installChar⚠

Rf_installDDVAL⚠

Rf_installS3Signature⚠

Rf_isArray⚠

Rf_isBasicClass⚠

Rf_isBlankString⚠

Rf_isComplex⚠

Rf_isEnvironment⚠

Rf_isExpression⚠

Rf_isFactor⚠

Rf_isFrame⚠

Rf_isFree⚠

Rf_isFunction⚠

Rf_isInteger⚠

Rf_isLanguage⚠

Rf_isList⚠

Rf_isLogical⚠

Rf_isMatrix⚠

Rf_isNewList⚠

Rf_isNull⚠

Rf_isNumber⚠

Rf_isNumeric⚠

Rf_isObject⚠

Rf_isOrdered⚠

Rf_isPairList⚠

Rf_isPrimitive⚠

Rf_isReal⚠

Rf_isS4⚠

Rf_isString⚠

Rf_isSymbol⚠

Rf_isTs⚠

Rf_isUnordered⚠

Rf_isUnsorted⚠

Rf_isUserBinop⚠

Rf_isValidString⚠

Rf_isValidStringF⚠

Rf_isVector⚠

Rf_isVectorAtomic⚠

Rf_isVectorList⚠

Rf_isVectorizable⚠

Rf_jump_to_toplevel⚠

Rf_lang1⚠

Rf_lang2⚠

Rf_lang3⚠

Rf_lang4⚠

Rf_lang5⚠

Rf_lang6⚠

Rf_lastElt⚠

Rf_lazy_duplicate⚠

Rf_lcons⚠

Rf_length⚠

Rf_lengthgets⚠

Rf_list1⚠

Rf_list2⚠

Rf_list3⚠

Rf_list4⚠

Rf_list5⚠

Rf_listAppend⚠

Rf_mainloop⚠

Rf_match⚠

Rf_matchE⚠

Rf_mkChar⚠

Rf_mkCharCE⚠

Rf_mkCharLen⚠

Rf_mkCharLenCE⚠

Rf_mkNamed⚠

Rf_mkString⚠

Rf_namesgets⚠

Rf_ncols⚠

Rf_nlevels⚠

Rf_nrows⚠

Rf_nthcdr⚠

Rf_onintr⚠

Rf_pmatch⚠

Rf_protect⚠

Rf_psmatch⚠

Rf_rPsort⚠

Rf_reEnc⚠

Rf_readS3VarsFromFrame⚠

Rf_revsort⚠

Rf_setAttrib⚠

Rf_setIVector⚠

Rf_setRVector⚠

Rf_setSVector⚠

Rf_setVar⚠

Rf_shallow_duplicate⚠

Rf_str2type⚠

Rf_stringPositionTr⚠

Rf_stringSuffix⚠

Rf_substitute⚠

Rf_translateChar⚠

Rf_translateChar0⚠

Rf_translateCharUTF8⚠

Rf_type2char⚠

Rf_type2rstr⚠

Rf_type2str⚠

Rf_type2str_nowarn⚠

Rf_unprotect⚠

Rf_unprotect_ptr⚠

Rf_warning⚠

Rf_warningcall⚠

Rf_warningcall_immediate⚠

Rf_xlength⚠

Rf_xlengthgets⚠

Rprintf⚠

Rvprintf⚠

SETCAD4R⚠

SETCADDDR⚠

SETCADDR⚠

SETCADR⚠

SETCAR⚠

SETCDR⚠

SETLENGTH⚠

SETLEVELS⚠

SET_ATTRIB⚠

SET_BODY⚠

SET_CLOENV⚠

SET_DDVAL⚠

SET_ENCLOS⚠

SET_ENVFLAGS⚠

SET_FORMALS⚠

SET_FRAME⚠

SET_HASHASH⚠

SET_HASHTAB⚠

SET_HASHVALUE⚠

SET_INTERNAL⚠

SET_MISSING⚠

SET_NAMED⚠

SET_OBJECT⚠

SET_PRCODE⚠

SET_PRENV⚠

SET_PRINTNAME⚠

SET_PRSEEN⚠

SET_PRVALUE⚠

SET_RDEBUG⚠

SET_RSTEP⚠

SET_RTRACE⚠

SET_S4_OBJECT⚠

SET_STRING_ELT⚠

SET_SYMVALUE⚠

SET_TAG⚠

SET_TRUELENGTH⚠

SET_TYPEOF⚠

SET_VECTOR_ELT⚠

STRING_ELT⚠

STRING_PTR⚠

SYMVALUE⚠

S_alloc⚠

S_realloc⚠

TAG⚠

TRUELENGTH⚠

TYPEOF⚠

UNIMPLEMENTED⚠

UNSET_S4_OBJECT⚠

VECTOR_ELT⚠

VECTOR_PTR⚠

WrongArgCount⚠

XLENGTH⚠

XTRUELENGTH⚠

_IO_feof⚠

_IO_ferror⚠

_IO_flockfile⚠

_IO_free_backup_area⚠

_IO_ftrylockfile⚠

_IO_funlockfile⚠

_IO_getc⚠

_IO_padn⚠

_IO_peekc_locked⚠

_IO_putc⚠

_IO_seekoff⚠

_IO_seekpos⚠

_IO_sgetn⚠

_IO_vfprintf⚠

_IO_vfscanf⚠

__acos⚠

__acosf⚠

__acosh⚠

__acoshf⚠

__acoshl⚠

__acosl⚠

__asin⚠

__asinf⚠

__asinh⚠

__asinhf⚠

__asinhl⚠

__asinl⚠

__atan⚠

__atan2⚠

__atan2f⚠

__atan2l⚠

__atanf⚠

__atanh⚠

__atanhf⚠

__atanhl⚠

__atanl⚠

__cbrt⚠

__cbrtf⚠

__cbrtl⚠

__ceil⚠

__ceilf⚠

__ceill⚠

__copysign⚠

__copysignf⚠

__copysignl⚠

__cos⚠

__cosf⚠

__cosh⚠

__coshf⚠

__coshl⚠

__cosl⚠

__drem⚠

__dremf⚠

__dreml⚠

__erf⚠

__erfc⚠

__erfcf⚠

__erfcl⚠

__erff⚠

__erfl⚠

__exp⚠

__exp2⚠

__exp2f⚠

__exp2l⚠

__expf⚠

__expl⚠

__expm1⚠

__expm1f⚠

__expm1l⚠

__fabs⚠

__fabsf⚠

__fabsl⚠

__fdim⚠

__fdimf⚠

__fdiml⚠

__finite⚠

__finitef⚠

__finitel⚠

__floor⚠

__floorf⚠

__floorl⚠

__fma⚠

__fmaf⚠

__fmal⚠

__fmax⚠

__fmaxf⚠

__fmaxl⚠

__fmin⚠

__fminf⚠

__fminl⚠

__fmod⚠

__fmodf⚠

__fmodl⚠

__fpclassify⚠

__fpclassifyf⚠

__fpclassifyl⚠

__frexp⚠

__frexpf⚠

__frexpl⚠

__gamma⚠

__gammaf⚠

__gammal⚠

__getdelim⚠

__hypot⚠

__hypotf⚠

__hypotl⚠

__ilogb⚠

__ilogbf⚠

__ilogbl⚠

__isinf⚠

__isinff⚠

__isinfl⚠

__isnan⚠

__isnanf⚠

__isnanl⚠

__j0⚠

__j0f⚠

__j0l⚠

__j1⚠

__j1f⚠

__j1l⚠

__jn⚠

__jnf⚠

__jnl⚠

__ldexp⚠

__ldexpf⚠

__ldexpl⚠

__lgamma⚠

__lgamma_r⚠

__lgammaf⚠

__lgammaf_r⚠

__lgammal⚠

__lgammal_r⚠

__llrint⚠

__llrintf⚠

__llrintl⚠

__llround⚠

__llroundf⚠

__llroundl⚠

__log⚠

__log1p⚠

__log1pf⚠

__log1pl⚠

__log2⚠

__log2f⚠

__log2l⚠

__log10⚠

__log10f⚠

__log10l⚠

__logb⚠

__logbf⚠

__logbl⚠

__logf⚠

__logl⚠

__lrint⚠

__lrintf⚠

__lrintl⚠

__lround⚠

__lroundf⚠

__lroundl⚠

__modf⚠

__modff⚠

__modfl⚠

__nan⚠

__nanf⚠

__nanl⚠

__nearbyint⚠

__nearbyintf⚠

__nearbyintl⚠

__nextafter⚠

__nextafterf⚠

__nextafterl⚠

__nexttoward⚠

__nexttowardf⚠

__nexttowardl⚠

__overflow⚠

__pow⚠

__powf⚠

__powl⚠

__remainder⚠

__remainderf⚠

__remainderl⚠

__remquo⚠

__remquof⚠

__remquol⚠

__rint⚠

__rintf⚠

__rintl⚠

__round⚠

__roundf⚠

__roundl⚠

__scalb⚠

__scalbf⚠

__scalbl⚠

__scalbln⚠

__scalblnf⚠

__scalblnl⚠

__scalbn⚠

__scalbnf⚠

__scalbnl⚠

__signbit⚠

__signbitf⚠

__signbitl⚠

__significand⚠

__significandf⚠

__significandl⚠

__sin⚠

__sinf⚠

__sinh⚠

__sinhf⚠

__sinhl⚠

__sinl⚠

__sqrt⚠

__sqrtf⚠

__sqrtl⚠

__tan⚠

__tanf⚠

__tanh⚠

__tanhf⚠

__tanhl⚠

__tanl⚠

__tgamma⚠

__tgammaf⚠

__tgammal⚠

__trunc⚠

__truncf⚠

__truncl⚠

__uflow⚠

__underflow⚠

__y0⚠

__y0f⚠

__y0l⚠

__y1⚠

__y1f⚠

__y1l⚠

__yn⚠

__ynf⚠

__ynl⚠

acos⚠

acosf⚠

acosh⚠

acoshf⚠

acoshl⚠

acosl⚠

asin⚠

asinf⚠

asinh⚠

asinhf⚠

asinhl⚠

asinl⚠

atan⚠

atan2⚠

atan2f⚠

atan2l⚠

atanf⚠

atanh⚠

atanhf⚠

atanhl⚠

atanl⚠

cbrt⚠

cbrtf⚠

cbrtl⚠

ceil⚠

ceilf⚠

ceill⚠

clearerr⚠

clearerr_unlocked⚠

copysign⚠

copysignf⚠

copysignl⚠

cos⚠

cosf⚠

cosh⚠

coshf⚠

coshl⚠

cosl⚠

ctermid⚠

dprintf⚠

drem⚠

dremf⚠

dreml⚠

erf⚠

erfc⚠

erfcf⚠

erfcl⚠

erff⚠

erfl⚠

exp⚠

exp2⚠

exp2f⚠

exp2l⚠

expf⚠

expl⚠

expm1⚠

expm1f⚠

expm1l⚠

fabs⚠

fabsf⚠

fabsl⚠

fclose⚠

fdim⚠

fdimf⚠

fdiml⚠

fdopen⚠

feof⚠

feof_unlocked⚠

ferror⚠

ferror_unlocked⚠

fflush⚠

fflush_unlocked⚠

fgetc⚠

fgetc_unlocked⚠

fgetpos⚠

fgets⚠

fileno⚠

fileno_unlocked⚠

findInterval⚠

find_interv_vec⚠

finite⚠

finitef⚠

finitel⚠

flockfile⚠

floor⚠

floorf⚠

floorl⚠

fma⚠

fmaf⚠

fmal⚠

fmax⚠

fmaxf⚠

fmaxl⚠

fmemopen⚠

fmin⚠

fminf⚠

fminl⚠

fmod⚠

fmodf⚠

fmodl⚠

fopen⚠

fprintf⚠

fpu_setup⚠

fputc⚠

fputc_unlocked⚠

fputs⚠

fread⚠

fread_unlocked⚠

freopen⚠

frexp⚠

frexpf⚠

frexpl⚠

fscanf⚠

fscanf1⚠

fseek⚠

fseeko⚠

fsetpos⚠

ftell⚠

ftello⚠

ftrylockfile⚠

funlockfile⚠

fwrite⚠

fwrite_unlocked⚠

gamma⚠

gammaf⚠

gammal⚠

getc⚠

getc_unlocked⚠

getchar⚠

getchar_unlocked⚠

getdelim⚠

getline⚠

getw⚠

hypot⚠

hypotf⚠

hypotl⚠

ilogb⚠

ilogbf⚠

ilogbl⚠

isinf⚠

isinff⚠

isinfl⚠

isnan⚠

isnanf⚠

isnanl⚠

j0⚠

j0f⚠

j0l⚠

j1⚠

j1f⚠

j1l⚠

jn⚠

jnf⚠

jnl⚠

ldexp⚠

ldexpf⚠

ldexpl⚠

lgamma⚠

lgamma_r⚠

lgammaf⚠

lgammaf_r⚠

lgammal⚠

lgammal_r⚠

llrint⚠

llrintf⚠

llrintl⚠

llround⚠

llroundf⚠

llroundl⚠

log⚠

log1p⚠

log1pf⚠

log1pl⚠

log2⚠

log2f⚠

log2l⚠

log10⚠

log10f⚠

log10l⚠

logb⚠

logbf⚠

logbl⚠

logf⚠

logl⚠

lrint⚠

lrintf⚠

lrintl⚠

lround⚠

lroundf⚠

lroundl⚠

matherr⚠

modf⚠

modff⚠

modfl⚠

nan⚠

nanf⚠

nanl⚠

nearbyint⚠

nearbyintf⚠

nearbyintl⚠

nextafter⚠

nextafterf⚠

nextafterl⚠

nexttoward⚠

nexttowardf⚠

nexttowardl⚠

open_memstream⚠

pclose⚠

perror⚠

popen⚠

pow⚠

powf⚠

powl⚠

printf⚠

process_site_Renviron⚠

process_system_Renviron⚠

process_user_Renviron⚠

putc⚠

putc_unlocked⚠

putchar⚠

putchar_unlocked⚠

puts⚠

putw⚠

remainder⚠

remainderf⚠

remainderl⚠

remove⚠

remquo⚠

remquof⚠

remquol⚠

rename⚠

renameat⚠

rewind⚠

rint⚠

rintf⚠

rintl⚠

round⚠

roundf⚠

roundl⚠

rsort_with_index⚠

scalb⚠

scalbf⚠

scalbl⚠

scalbln⚠

scalblnf⚠

scalblnl⚠

scalbn⚠

scalbnf⚠

scalbnl⚠

scanf⚠

scanf1⚠

setbuf⚠

setbuffer⚠

setlinebuf⚠

setup_Rmainloop⚠

setvbuf⚠

significand⚠

significandf⚠

significandl⚠

sin⚠

sinf⚠

sinh⚠

sinhf⚠

sinhl⚠

sinl⚠

snprintf⚠

sprintf⚠

sqrt⚠

sqrtf⚠

sqrtl⚠

sscanf⚠

sscanf1⚠

tan⚠

tanf⚠

tanh⚠

tanhf⚠

tanhl⚠

tanl⚠

tempnam⚠

tgamma⚠

tgammaf⚠

tgammal⚠

tmpfile⚠

tmpnam⚠

tmpnam_r⚠

trunc⚠

truncf⚠

truncl⚠

ungetc⚠

vdprintf⚠

vfprintf⚠

vfscanf⚠

vfscanf1⚠

vmaxget⚠

vmaxset⚠

vprintf⚠

vscanf⚠

vscanf1⚠

vsnprintf⚠

vsprintf⚠

vsscanf⚠

vsscanf1⚠

y0⚠

y0f⚠

y0l⚠

y1⚠

y1f⚠

y1l⚠

yn⚠

ynf⚠

ynl⚠

Type Definitions

FILE

PROTECT_INDEX

ParseStatus

R_CFinalizer_t

R_allocator_t

R_inpstream_t

R_len_t

R_outpstream_t

R_pstream_data_t

R_pstream_format_t

R_xlen_t

Rboolean

Rbyte

Rstart

SA_TYPE

SEXP

SEXPTYPE

_IO_lock_t

_LIB_VERSION_TYPE

__FILE

__blkcnt64_t

__blkcnt_t

__blksize_t

__builtin_va_list

__caddr_t

__clock_t

__clockid_t

__codecvt_result

__daddr_t

__dev_t

__fsblkcnt64_t

__fsblkcnt_t

__fsfilcnt64_t

__fsfilcnt_t

__fsword_t

__gid_t

__gnuc_va_list

__id_t

__ino64_t

__ino_t

__int8_t

__int16_t

__int32_t

__int64_t

__intptr_t

__io_close_fn

__io_read_fn

__io_seek_fn

__io_write_fn

__key_t

__loff_t

__mode_t

__nlink_t

__off64_t

__off_t

__pid_t

__qaddr_t

__quad_t

__rlim64_t

__rlim_t

__socklen_t

__ssize_t

__suseconds_t

__syscall_slong_t

__syscall_ulong_t

__time_t

__timer_t

__u_char

__u_int

__u_long

__u_quad_t

__u_short

__uid_t

__uint8_t

__uint16_t

__uint32_t

__uint64_t

__useconds_t

_bindgen_ty_1

cetype_t

double_t

float_t

fpos_t

nchar_type

off_t

size_t

ssize_t

va_list

wchar_t

Unions

__mbstate_t__bindgen_ty_1